requirements.txt
asgiref==3.9.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
Deprecated==1.2.18
dnspython==2.7.0
dotenv==0.9.9
Flask==3.1.1
Flask-PyMongo==3.0.1
googleapis-common-protos==1.59.1
grpcio==1.74.0
idna==3.10
importlib-metadata==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
opentelemetry-api==1.24.0
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-asyncio==0.45b0
opentelemetry-instrumentation-aws-lambda==0.45b0
opentelemetry-instrumentation-dbapi==0.45b0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-grpc==0.45b0
opentelemetry-instrumentation-jinja2==0.45b0
opentelemetry-instrumentation-logging==0.45b0
opentelemetry-instrumentation-pymongo==0.45b0
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-sqlite3==0.45b0
opentelemetry-instrumentation-urllib==0.45b0
opentelemetry-instrumentation-urllib3==0.45b0
opentelemetry-instrumentation-wsgi==0.45b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-test-utils==0.45b0
opentelemetry-util-http==0.45b0
opentelemetry-distro[otlp]==0.45b0
packaging==25.0
protobuf==4.25.8
pymongo==4.14.0
python-dotenv==1.1.1
requests==2.32.4
thrift==0.22.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
